chore: upgrade to Django 4.2
